requirements.txt
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-documentai==2.15.0
googleapis-common-protos==1.59.0
pandas==2.0.0